requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-migrate==4.0.7
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.0
requests==2.32.3
urllib3==1.26.20
waitress==2.1.2
Werkzeug==3.0.4
Flask-SQLAlchemy==3.1.1
typing-extensions==4.12.2
greenlet==3.0.1
APScheduler==3.10.4
Flask-Mail==0.10.0
cloudmersive-virus-api-client==3.0.1
cloudmersive-validate-api-client==3.2.5
boto3==1.35.46
botocore==1.35.46
xhtml2pdf==0.2.11
flask_session==0.8.0
